// pages/Search/interfaces.ts
export interface PropertiesTabPanelFields {
  propertyNumber?: string | undefined;
  city?: string | undefined;
  postCode?: string | undefined;
  street?: string | undefined;
  houseNumber?: string | undefined;
  partnerCode?: string | undefined;
}

export interface ConsumptionUnitTabPanelFields {
  residentialUnit?: string | undefined;
  userName?: string | undefined;
  propertyNumber?: string | undefined;
}